requirements.txt
pandas
numpy
scipy
requests
python-dotenv
gspread
google-auth
xlsxwriter
